Cambio en la estructura del proyecto, mas que oanda en los actions y reducers. Tambien modifique el header para que herede de props la funcion logout
