+++ src/store/records/recordsConstants.js
@@ -1,13 +1,3 @@
-export const LOGOUT = 'LOGOUT';
-
-export const LOGIN_OR_REGISTER_PENDING = 'LOGIN_OR_REGISTER_PENDING';
-export const LOGIN_OR_REGISTER_SUCCESS = 'LOGIN_OR_REGISTER_SUCCESS';
-export const LOGIN_OR_REGISTER_FAILED = 'LOGIN_OR_REGISTER_FAILED';
-
-export const RESET_PASSWORD_PENDING = 'RESET_PASSWORD_PENDING';
-export const RESET_PASSWORD_SUCCESS = 'RESET_PASSWORD_SUCCESS';
-export const RESET_PASSWORD_FAILED = 'RESET_PASSWORD_FAILED';
-
 export const ADD_RECORD_PENDING = 'ADD_RECORD_PENDING';
 export const ADD_RECORD_SUCCESS = 'ADD_RECORD_SUCCESS';
 export const ADD_RECORD_FAILED = 'ADD_RECORD_FAILED';
